pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chunkrag"
version = "0.1.0"
description = "Course project comparing chunking strategies for retrieval-augmented QA."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "accelerate>=0.34,<1",
  "datasets>=2.20,<3",
  "faiss-cpu>=1.8,<2",
  "langchain-text-splitters>=0.2.2,<0.4",
  "numpy>=1.26,<2",
  "pandas>=2.2,<3",
  "sentence-transformers>=3,<4",
  "sentencepiece>=0.2,<1",
  "spacy>=3.7,<4",
  "torch>=2.3,<3",
  "transformers>=4.45,<5",
  "tqdm>=4.66,<5",
]

[project.optional-dependencies]
chonkie = [
  "chonkie[semantic]>=1,<2",
]

[tool.setuptools.packages.find]
where = ["src"]
